+++ app/assets/javascripts/google_cloud/aiml/panel.vue
@@ -1,12 +1,10 @@
 <!-- eslint-disable vue/multi-word-component-names -->
 <script>
 import GoogleCloudMenu from '../components/google_cloud_menu.vue';
-import IncubationBanner from '../components/incubation_banner.vue';
 import ServiceTable from './service_table.vue';
 
 export default {
   components: {
-    IncubationBanner,
     GoogleCloudMenu,
     ServiceTable,
   },
@@ -45,8 +43,6 @@ export default {
 
 <template>
   <div>
-    <incubation-banner />
-
     <google-cloud-menu
       active="aiml"
       :configuration-url="configurationUrl"