+++ doc/user/project/merge_requests/changes.md
@@ -254,7 +254,7 @@ To change how a merge request shows changed lines:
 
 {{< collapsible title="Model information" >}}
 
-- Default LLM: Anthropic [Claude Sonnet 4.5](https://console.cloud.google.com/vertex-ai/publishers/anthropic/model-garden/claude-sonnet-4-5)
+- [Default LLM](../../gitlab_duo/model_selection.md#default-models)
 - LLM for Amazon Q: Amazon Q Developer
 
 {{< /collapsible >}}